Moved utils to catalog-utils file
