+++ frontend/packages/dev-console/src/components/catalog/CatalogPage.tsx
@@ -10,7 +10,7 @@ import {
   useCatalogCategories,
   CatalogController,
 } from '@console/shared';
-import { isCatalogTypeEnabled } from '../../utils/useAddActionExtensions';
+import { isCatalogTypeEnabled } from '../../utils/catalog-utils';
 import NamespacedPage, { NamespacedPageVariants } from '../NamespacedPage';
 import CreateProjectListPage from '../projects/CreateProjectListPage';
